rptest: separate LocalPayloadDir from omb service
